# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "operationId": "CustomerController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a customer",
        "tags": [
          "customers"
        ]
      },
      "get": {
        "operationId": "CustomerController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List customers",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{customerId}/contact": {
      "post": {
        "operationId": "CustomerController_createContact_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request or customer ID is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create contact",
        "tags": [
          "customers"
        ]
      },
      "get": {
        "operationId": "CustomerController_findContact_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get contact",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomerController_updateContact_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update contact",
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "CustomerController_deleteContact_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete contact",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{customerId}": {
      "get": {
        "operationId": "CustomerController_findById_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "The customer ID is not a UUIDv7."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a customer",
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomerController_update_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a customer",
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "CustomerController_delete_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a customer",
        "tags": [
          "customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Faberon API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Resource was not found."
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "CreateContactRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "ada@example.com",
            "maxLength": 320,
            "nullable": true
          },
          "phone": {
            "type": "string",
            "example": "+4798765432",
            "maxLength": 32,
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Primary contact",
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCustomerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme AS",
            "minLength": 1,
            "maxLength": 120
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "company",
              "government"
            ],
            "example": "company",
            "description": "Defaults to company when omitted."
          },
          "legalName": {
            "type": "string",
            "example": "Acme AS",
            "maxLength": 200,
            "nullable": true
          },
          "taxNumber": {
            "type": "string",
            "example": "NO999888777",
            "maxLength": 64,
            "nullable": true
          },
          "address": {
            "type": "string",
            "example": "Karl Johans gate 1",
            "maxLength": 240,
            "nullable": true
          },
          "city": {
            "type": "string",
            "example": "Oslo",
            "maxLength": 120,
            "nullable": true
          },
          "region": {
            "type": "string",
            "example": "Oslo",
            "maxLength": 120,
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "example": "0154",
            "maxLength": 32,
            "nullable": true
          },
          "country": {
            "type": "string",
            "enum": [
              "pl",
              "no"
            ],
            "example": "no",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "example": "Preferred customer",
            "nullable": true
          },
          "contact": {
            "$ref": "#/components/schemas/CreateContactRequest"
          }
        },
        "required": [
          "name"
        ]
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "019535d9-3df7-79fb-b466-fa907fa17f9e"
          },
          "companyId": {
            "type": "string",
            "format": "uuid",
            "example": "019535d9-3df7-79fb-b466-fa907fa17f9e"
          },
          "name": {
            "type": "string",
            "example": "Acme AS"
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "company",
              "government"
            ]
          },
          "legalName": {
            "type": "string",
            "nullable": true
          },
          "taxNumber": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "enum": [
              "pl",
              "no"
            ],
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "companyId",
          "name",
          "type",
          "legalName",
          "taxNumber",
          "address",
          "city",
          "region",
          "postalCode",
          "country",
          "notes"
        ]
      },
      "ContactResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "019535d9-3df8-7abc-8def-fa907fa17f9f"
          },
          "customerId": {
            "type": "string",
            "format": "uuid",
            "example": "019535d9-3df7-79fb-b466-fa907fa17f9e"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "customerId",
          "name",
          "email",
          "phone",
          "description"
        ]
      },
      "UpdateCustomerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme AS",
            "minLength": 1,
            "maxLength": 120
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "company",
              "government"
            ],
            "example": "company"
          },
          "legalName": {
            "type": "string",
            "example": "Acme AS",
            "maxLength": 200,
            "nullable": true
          },
          "taxNumber": {
            "type": "string",
            "example": "NO999888777",
            "maxLength": 64,
            "nullable": true
          },
          "address": {
            "type": "string",
            "example": "Karl Johans gate 1",
            "maxLength": 240,
            "nullable": true
          },
          "city": {
            "type": "string",
            "example": "Oslo",
            "maxLength": 120,
            "nullable": true
          },
          "region": {
            "type": "string",
            "example": "Oslo",
            "maxLength": 120,
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "example": "0154",
            "maxLength": 32,
            "nullable": true
          },
          "country": {
            "type": "string",
            "enum": [
              "pl",
              "no"
            ],
            "example": "no",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "example": "Preferred customer",
            "nullable": true
          }
        }
      },
      "UpdateContactRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ada Lovelace",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "ada@example.com",
            "maxLength": 320,
            "nullable": true
          },
          "phone": {
            "type": "string",
            "example": "+4798765432",
            "maxLength": 32,
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "Primary contact",
            "nullable": true
          }
        }
      },
...
    }
  }
}